requirements.txt
asgiref==3.2.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.2
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.9.0
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
gunicorn==20.0.4
idna==2.8
imageio==2.8.0
imageio-ffmpeg==0.4.1
moviepy==1.0.1
mysqlclient==1.4.6
numpy==1.18.1
Pillow==7.0.0
proglog==0.1.9
protobuf==3.11.3
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
tqdm==4.43.0
urllib3==1.25.7
whitenoise==5.0.1
